+++ messages/en.json
@@ -2737,16 +2737,24 @@
     "group_bookkeeping": "Bookkeeping",
     "group_payroll": "Payroll",
     "group_pending_operations": "Staged operations",
-    "group_documents": "Documents (REST API)",
-    "group_companies": "Companies (REST API)",
-    "group_events": "Events (REST API)",
-    "group_webhooks": "Webhooks (REST API)",
-    "group_operations": "Operations (REST API)",
-    "group_compliance": "Compliance (REST API)",
+    "group_documents": "Documents",
+    "group_companies": "Companies",
+    "group_events": "Events",
+    "group_webhooks": "Webhooks",
+    "group_operations": "Operations",
+    "group_compliance": "Compliance",
+    "group_articles": "Articles",
+    "group_reconciliation": "Reconciliation",
+    "group_rest_only": "{name} (REST API)",
     "scope_transactions_read": "Read: list transactions, templates, categories, inbox",
     "scope_transactions_write": "Write: categorize, receipt matching, link to invoice, document upload",
+    "scope_reconciliation_read": "Read: accounts to reconcile, the bridge per account and the items behind it (bank + tax account)",
+    "scope_reconciliation_write": "Write: match and unmatch events against vouchers, ignore items (staged)",
+    "scope_reconciliation_signoff": "Sign off: mark an account as reconciled through a date, or reopen a sign-off (staged)",
     "scope_customers_read": "Read: list customers",
     "scope_customers_write": "Write: create customers",
+    "scope_articles_read": "Read: list articles in the article register",
+    "scope_articles_write": "Write: create and update articles",
     "scope_invoices_read": "Read: list invoices",
     "scope_invoices_write": "Write: create, send, mark paid/sent, credit, convert",
     "scope_suppliers_read": "Read: list suppliers and supplier invoices",
@@ -2760,6 +2768,7 @@
     "scope_documents_read": "Read: list and fetch document attachments",
     "scope_documents_write": "Write: upload and link documents to vouchers",
     "scope_companies_read": "Read: company profiles the key has access to",
+    "scope_companies_write": "Write: create companies and update company settings",
     "scope_events_read": "Read: poll event_log as a webhook fallback",
     "scope_webhooks_manage": "Manage: create, list, update, delete subscriptions",
     "scope_operations_read": "Read: status of long-running operations (import, year-end, revaluation)",
