+++ messages/sv.json
@@ -2737,16 +2737,24 @@
     "group_bookkeeping": "Bokföring",
     "group_payroll": "Löner",
     "group_pending_operations": "Stagade operationer",
-    "group_documents": "Dokument (REST API)",
-    "group_companies": "Företag (REST API)",
-    "group_events": "Händelser (REST API)",
-    "group_webhooks": "Webhooks (REST API)",
-    "group_operations": "Operationer (REST API)",
-    "group_compliance": "Compliance (REST API)",
+    "group_documents": "Dokument",
+    "group_companies": "Företag",
+    "group_events": "Händelser",
+    "group_webhooks": "Webhooks",
+    "group_operations": "Operationer",
+    "group_compliance": "Compliance",
+    "group_articles": "Artiklar",
+    "group_reconciliation": "Avstämning",
+    "group_rest_only": "{name} (REST API)",
     "scope_transactions_read": "Läs: lista transaktioner, mallar, kategorier, inbox",
     "scope_transactions_write": "Skriv: kategorisera, kvittomatchning, koppling mot faktura, dokumentuppladdning",
+    "scope_reconciliation_read": "Läs: konton att stämma av, bryggan per konto och raderna bakom den (bank + skattekonto)",
+    "scope_reconciliation_write": "Skriv: koppla och koppla bort händelser mot verifikat, ignorera rader (stagas)",
+    "scope_reconciliation_signoff": "Signera: markera ett konto som avstämt t.o.m. ett datum, eller öppna signeringen igen (stagas)",
     "scope_customers_read": "Läs: lista kunder",
     "scope_customers_write": "Skriv: skapa kunder",
+    "scope_articles_read": "Läs: lista artiklar i artikelregistret",
+    "scope_articles_write": "Skriv: skapa och uppdatera artiklar",
     "scope_invoices_read": "Läs: lista fakturor",
     "scope_invoices_write": "Skriv: skapa, skicka, markera betald/skickad, kreditera, konvertera",
     "scope_suppliers_read": "Läs: lista leverantörer och leverantörsfakturor",
@@ -2760,6 +2768,7 @@
     "scope_documents_read": "Läs: lista och hämta dokumentbilagor",
     "scope_documents_write": "Skriv: ladda upp och koppla dokument till verifikationer",
     "scope_companies_read": "Läs: företagsprofiler nyckeln har åtkomst till",
+    "scope_companies_write": "Skriv: skapa företag och uppdatera företagsinställningar",
     "scope_events_read": "Läs: polla event_log som webhook-fallback",
     "scope_webhooks_manage": "Hantera: skapa, lista, uppdatera, radera prenumerationer",
     "scope_operations_read": "Läs: status för långkörande operationer (import, bokslut, omvärdering)",
